Move download_emoji.go to subcmd
